# [NETBEANS-4901] Remove the `--progressbar` option