@junichi11 2020-10-30
@lkishalmi PHPDocumentor 3.0.0 was already released: https://github.com/phpDocumentor/phpDocumentor/releases/tag/v3.0.0
Can we include this to 12.2 if it's OK with @tmysik?